Handling duplicate title name and rejecting them
Duplicate names produce issues when it comes to saving documents with the same name to the file system.

Now we handle this properly but not allowing to save documents that have the same name on the app.

There was also some minor refactoring of the save component, leaving only one invokation an action, transfering the rest of the logic to the state, where it is suposed to live.